// src/types/exam-results.ts
export interface IExamResult {
  year: string;
  exam: string;
  totalExaminee: number;
  totalPassed: number;
  totalFailed: number;
  totalPassPercentage: number;
  totalFailPercentage: number;
  totalPassMale: number;
  totalPassFemale: number;
  totalAPlus: number;
  totalAGrade: number;
  totalAMinus: number;
  totalB: number;
  totalC: number;
  totalD: number;
  _id: string;
  createdAt: string;
  updatedAt: string;
}